Random suffix didn't work (I think helm is using thr same seed on each resync?). Could use a timestamp but makes more sense I think to rerun the job iff config changes.
Any change to instance config will trigger a rerun of the job. We can refine this in future to only take into account a subset of instance config (maybe just the ibm_sls key?). But the job is idempotent and quick so no real harm in running it when we don't really need to.

Also: use a static job label to avoid resyncing networking policy unnecessarily.