Update 04-postsync-maximoit-verify.yaml
Issue
Test cases were failing in the production environment due to the absence of demo data. Specifically, the maxadmin user lacked the necessary privileges to access the maxapps object structure, which was being used in our API request. As a result, the API call to mxapimaxapp was returning insufficient or no data, causing the tests to fail.

Fix
To resolve this, we’ve updated the API request to use the mxapigroup object instead. This object is accessible to the maxadmin user even in the absence of demo data, ensuring consistent behavior across environments.

Changes Made
Replaced the API endpoint: From:

{MANAGE_URL}/maximo/api/os/mxapimaxapp?oslc.select=*
To:

{MANAGE_URL}/maximo/api/os/mxapigroup?lean=1&oslc.select=*&oslc.where=groupname='Maxadmin'